rework tests a bit. Now it is clear
